% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-med.R
\name{fit.mp}
\alias{fit.mp}
\title{Fit Mplus model}
\usage{
fit.mp(
  l.mo,
  m.b = "no_confound",
  v.nm.dv = v.nm.dv,
  v.nm.iv = v.nm.iv,
  int.cores = parallel::detectCores(),
  s.iter = 10000L
)
}
\description{
Fit Mplus model
}
\keyword{internal}
