% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo-med.R
\name{gen.mo.par}
\alias{gen.mo.par}
\title{Generate model parameters for causal effects}
\usage{
gen.mo.par(l.mo, v.nm.dv, v.nm.iv, b.exv)
}
\description{
Generate model parameters for causal effects
}
\keyword{internal}
