% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-med-sa.R
\name{gen.sa.testseq}
\alias{gen.sa.testseq}
\title{Generate sensitivity analysis test sequence}
\usage{
gen.sa.testseq(s.mcor, s.dec)
}
\description{
Generate sensitivity analysis test sequence
}
\keyword{internal}
