% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-med.R
\name{gen.sim.par}
\alias{gen.sim.par}
\title{Generate simulation parameters}
\usage{
gen.sim.par(s.seed, l.mo, l.mo.par, b.ev, m.b, s.meth, s.mode, int.cores)
}
\description{
Generate simulation parameters
}
\keyword{internal}
