% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gen_model}
\alias{gen_model}
\title{Generate model from data generation parameters}
\usage{
gen_model(n, s.seed, l.static, l.coef, v.norm.sd, s.dat)
}
\description{
Generate model from data generation parameters
}
\keyword{internal}
