% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-med-sa.R
\name{get.sa.maxcor.allsteps}
\alias{get.sa.maxcor.allsteps}
\title{Get maximum correlation (all steps)}
\usage{
get.sa.maxcor.allsteps(
  s.dec,
  m.u.b,
  m.res.var,
  v.res.sdp,
  v.u.dir,
  l.mo,
  v.nm.dv,
  v.nm.iv
)
}
\description{
Get maximum correlation (all steps)
}
\keyword{internal}
