% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-med-sa.R
\name{get.sa.maxcor.step}
\alias{get.sa.maxcor.step}
\title{Get maximum correlation (single step)}
\usage{
get.sa.maxcor.step(
  v.step.int,
  s.cor.start,
  m.u.b,
  m.res.var,
  v.res.sdp,
  v.u.dir,
  l.mo,
  v.nm.dv,
  v.nm.iv
)
}
\description{
Get maximum correlation (single step)
}
\keyword{internal}
