% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo-med.R
\name{mo.1dep}
\alias{mo.1dep}
\title{Model counterfactuals (1+ dependencies)}
\usage{
mo.1dep(l.set.x, l.mo.cf, s.nm.mo, l.eff.cft, l.eff.cfg, l.dep, l.fam, b.exv)
}
\description{
Model counterfactuals (1+ dependencies)
}
\keyword{internal}
