% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo-med.R
\name{mo.ce.est}
\alias{mo.ce.est}
\title{Model-based causal effect estimates}
\usage{
mo.ce.est(
  l.mo,
  l.fam,
  v.nm.dv,
  v.nm.iv,
  b.exv,
  l.dep,
  l.eff.cfg,
  l.eff.cft,
  l.eff.def.split
)
}
\description{
Model-based causal effect estimates
}
\keyword{internal}
