% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo-med.R
\name{mo.cf.est}
\alias{mo.cf.est}
\title{Model-based counterfactual estimates}
\usage{
mo.cf.est(
  l.mo,
  l.fam,
  v.nm.dv,
  v.nm.iv,
  b.exv,
  l.dep,
  l.eff.cfg,
  l.eff.cft,
  b.mn = TRUE
)
}
\description{
Model-based counterfactual estimates
}
\keyword{internal}
