% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup.env}
\alias{setup.env}
\title{Setup consistent environment across functions}
\usage{
setup.env(
  list.models,
  int.iter,
  b.parallel = "auto",
  int.cores = "max",
  canonical = FALSE
)
}
\description{
Setup consistent environment across functions
}
\keyword{internal}
