% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-med.R
\name{sim.1dep}
\alias{sim.1dep}
\title{Simulation counterfactuals (1+ dependencies)}
\usage{
sim.1dep(l.set.x, l.sim.cf, s.nm.mo, l.eff.cf, l.eff.cfg, l.dep, l.fam, b.exv)
}
\description{
Simulation counterfactuals (1+ dependencies)
}
\keyword{internal}
