% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-med.R
\name{sim.cf}
\alias{sim.cf}
\title{Run simulation counterfactuals}
\usage{
sim.cf(l.mo, s.mode, m.b, s.cof.mth, int.sims, s.seed, b.parallel, int.cores)
}
\description{
Run simulation counterfactuals
}
\keyword{internal}
