% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-med.R
\name{sim.med.1}
\alias{sim.med.1}
\title{Single simulation run}
\usage{
sim.med.1(
  s.seed,
  l.mo,
  l.mo.par,
  l.eff.cf,
  l.eff.cfg,
  l.dep,
  b.exv,
  m.b,
  s.cof.mth,
  s.mode
)
}
\description{
Single simulation run
}
\keyword{internal}
